Skip 'is-ancestor' check for shallow clones with SHA revision
In Ie34c5d860, a strict check was introduced to ensure that the provided
SHA1 is an ancestor or equal to the tip-of-tree of the upstream ref, but
with shallow clones (especially with depth 1), this check will fail when
the upstream ref becomes stale. Skipping the is-ancestor check should be
OK with shallow clones as repo operations such as rebase will likely not
be attempted.

Change-Id: Ib3c0b71c7a1efc104458e094a664da8771e713fa