project: fix alternate path for subproject when use --reference
When creating repository mirrors, if the project has submodules,
they will be located at project.git/subprojects/submodule.git

Currently, when using a mirror with the --reference flag, the
submodule mirror is searched for at project/submodule.git. Since
there is no mirror at this path, the submodule starts downloading
again.

To fix this behavior, the calculation of the real path to the
mirror has been added for submodules.

Change-Id: I879e2cfb686ad1639f84b744a7fa4b8194b75e76
Signed-off-by: Yaroslav Borunov <yaborunov@salutedevices.com>