sync: add flag --ignore-upstream
If you use a manifest with commit hashes in revisions to download
code with a given depth and the commits are not head commits, then
when downloading an upstream branch, repo cannot always determine
that the commit in revision is part of upstream. Because of this,
the repository is downloaded with the full history of commits,
which significantly slows down the code download process.

The --ignore-upstream flag allows you to bypass the check and
download only the number of commits specified in the depth,
ignoring the download of upstream

This can be used for CI systems where upstream is not important and
only the source code from revisions for the build is needed.

Change-Id: I2819dfa24eeda7a989c0743478f5ba066ecbcee8
Signed-off-by: Yaroslav Borunov <yaborunov@salutedevices.com>