manifest: Add new subcommand for printing include tree
"repo includetree" is a new subcommand, that will print all included
manifests in a hierarchical tree-structure to standard output.
This is useful in larger manifest projects where there are often multiple
xml-files included from various suppliers and extend-projects used to add
patches on upstream projects.

To enable this change, the _ParseManifestXml function was split in
two, to reuse the common logic of validating valid include
names relative to the include_root, while not having nodes affected by
the attribute-mutation that happens when groups are inherited from the
parent <include>.

Change-Id: Id5c4e47a7be7035ba8e06931274b0095aecf6276