Fix FileNotFoundError in git superproject
I had a workspace. I do not remember if I used the --no-use-superproject
flag during the init. I did switch branches with "repo init -b foo"
after which I was unable to use repo upload as it would throw a
FileNotFoundError when attempting to find the superproject information.
Even after re-initializing the workspace with the proper
--no-use-superproject flag and confirming in .repo/manifests.git/config
that the superproject was set to false, the exception continued as the
manifest.sp returned an object in _UploadBranch of subcmds/uploads.py
but repo_id would blow up.
I was unable to find a way to get the repo tool to realize there should
be no superproject in my workspace so the easiest solution was to catch
the exception and return None to get the upload task to believe there is
no superproject.

Change-Id: I799a441069adf9ee928114159cf1aaabb21b9d4f
Signed-off-by: Travis Allen <talle112@ford.com>