sync: explicity fetch revision if sha1
Explicitly fetch the revision if sha1 in case it is not part of the
history of any upstream branch or tag.
Otherwise, if shallow cloning/depth isn't set, then sync will fail
later on when trying to check out the revision since it was never
fetched.
This covers the case where revision is set to an unmerged commit.

For this to work, the git remote must also support fetching
commits directly
For gerrit: uploadpack.allowReachableSHA1InWant would need
to be set:
https://github.com/git/git/blob/v2.5.0/Documentation/config.txt#L2570

Change-Id: I849fa746c954e8823ae869c2d7715dfd05c70d6d