RFC: Replace minidom with ElementTree
PRELUDE:
    This was just a rough MVP to measure if there
    was any improvement to be found.
    While not the big improvements i was hoping for,
    i'm still leaving this here to probe for if
    there is any interest in this and to share the numbers
    for the curious.
    If yes, i can continue to iron out the remaining
    details to avoid regressions.
/PRELUDE

The main benefits of this are
* Improve "repo manifest" performance with 200ms (30%)
* ElementTree is more popular and has a easier,
  more pythonic API to work with. In fact, the docs
  for minidom itself directs users towards ElementTree [0]
* As a side effect, I added type hints, which makes the
  code more readable and maintanable. This is however
  also possible with the old parser.

This may seem very negligble in the context of
"repo sync" taking several minutes. The reason i
ended up here was while developing a script to
analyze the history of a manifest repo, containing
thousands of commits. Running `repo manifest` turned
out to be biggest bottleneck, often taking over
half a second to run, totalling up to several minutes.

After profiling repo, manifest_xml.py was identified as
the main culprit, noticing it uses minidom, which is
known to be slower than ElementTree i decided to
try switching to ElementTree to see if it would improve
performance.

The results show a 100-200ms improvement,
measuring `repo manifest` containing
close to 2000 projects.
                 ET   minidom
    Mean:     255ms   349ms
    Min:      224ms   280ms
    Max:      314ms   548ms
    P90:      286ms   412ms
    P95:      297ms   497ms
    P99:      314ms   548ms

While not as big as an improvement as i hoped for,
it's still noticable under some scenarios, and also
other improvements.

The biggest bottleneck remaining is still xml serialization
within ElementTree, which is difficult to optimize
further. The ratio of time spent in load vs save is now also
shifted from 50/50 to 1/3 loading and 2/3 saving. Meaning that
operations that don't output the manifest benefit marginally
more.

Current known limitations
* The previous implementation would put a newline after
  some tags, such as <default> and <notice>.
  This is not yet implemented in this patchset.

The biggest caveats good to be aware of is
* minidom getAttribute returns empty strings on missing
  attribute, whereas ElementTree returns None.

[0] https://docs.python.org/3/library/xml.dom.minidom.html
Change-Id: Ib2d400e10f79a7c3fb294863092bfacb10f2d701

Change-Id: Id1aaafec331b1717ff846a8407163bacdb292aeb