Fix submodule initialization in interleaved sync mode
With the introduction of interleaved sync mode, the submodule
activation logic broke because the 'has_submodules' attribute
was no longer being populated when needed. With this change,
each submodule is initialized when it enters the Sync_LocalHalf
stage, whereas previously all submodules were initialized at
once when the parent repository entered the Sync_LocalHalf stage.

This change makes the submodule activation logic more robust and
less prone to breakage.

Change-Id: I25eca4ea2a6868219045cfa088988eb01ded47d2