Use 'git rebase' during 'repo sync --rebase'
'repo sync --rebase' should do a rebase if it encounters local commits
during a 'repo sync'.
This was broken by http://go/grev/437421, which caused this to execute
the '_doff' hook (which stands for 'do fast forward'), which is
implemented using 'git merge --no-stat'.

This caused *multiple* actual editor windows to pop up during
'repo sync --rebase', asking the user to enter a commit message for the
merge.

In this CL I explicitly make that code path do a 'git rebase'.

Bug: b/434565811
Test: Used it myself for over a week.
Change-Id: I0bf3ff181f15b9d5b2e3f85f7f84e302139fdab7