Add submodule-override support
In certain use cases, upstream repositories are mirrored internally,
and the fetch URL of the mirror is used in the manifest to work around
issues such as rate limiting or inconsistent fetch performance. However,
a limitation of repo is that if such a repository contains submodules,
there is no way to fetch those submodules from an internal mirror as
well.

This change introduces a new submodule-override element (which is
specified as a child of the project element) in the manifest format.
It allows specifying a custom URL for a submodule based on its path,
enabling consistent use of internal mirrors for both repositories
and their submodules.

Change-Id: Icd4d6ac91ed38287607bd10341fe0ce66017793e