Handle potential leftover bare gitdir from a failed sync
The gitdir for a project may be left in a state with bare=true due
to a previous failed sync. In this state, during a subsequent sync
attempt, repo will skip initializing the gitdir (since the directory
already exists) and directly attempt to checkout the worktree, which
will fail because the project is bare.

Bug: 457478027
Change-Id: I4767494a3a54e7734174eae3a0d939fa9d174288